requirements.txt
translatepy==2.3
python-telegram-bot==22.0
psycopg2==2.9.10
pandas==2.2.3
dotenv==0.9.9
pytest==8.3.5
httpx==0.27.0